More functions and nested loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 
 int main(void)
 {
-	unsigned int x1 = 1;
-	unsigned int x2 = 1;
-	unsigned int x3 = 2;
+	long int x1 = 1;
+	long int x2 = 1;
+	long int x3 = 2;
 	int counter = 0;
 
 	while (counter < 50)
 	{
-		printf("%u, ", x3);
+		printf("%ld, ", x3);
 		x3 = x1 + x2;
 		x1 = x2;
 		x2 = x3;
